// packages/domain/src/pages/schemas.ts
import { PageType } from '@repo/db'
import { z } from 'zod'

export const createPageInput = z.object({
  workspaceId: z.string().uuid(),
  parentId: z.string().uuid().nullable(),
  title: z.string().optional(),
  icon: z.string().optional(),
  type: z.nativeEnum(PageType).optional(),
})
export type CreatePageInput = z.infer<typeof createPageInput>

export const renamePageInput = z.object({
  id: z.string().uuid(),
  workspaceId: z.string().uuid(),
  title: z.string(),
  icon: z.string().nullable().optional(),
})
export type RenamePageInput = z.infer<typeof renamePageInput>

export const updatePageInput = z.object({
  id: z.string().uuid(),
  workspaceId: z.string().uuid(),
  title: z.string().optional(),
  icon: z.string().nullable().optional(),
  type: z.nativeEnum(PageType).optional(),
})
export type UpdatePageInput = z.infer<typeof updatePageInput>

export const duplicatePageInput = z.object({
  pageId: z.string().uuid(),
})
export type DuplicatePageInput = z.infer<typeof duplicatePageInput>

export const movePageInput = z.object({
  pageId: z.string().uuid(),
  newParentId: z.string().uuid().nullable(),
})
export type MovePageInput = z.infer<typeof movePageInput>

export const reorderPageInput = z.object({
  pageId: z.string().uuid(),
  newParentId: z.string().uuid().nullable(),
  newPrevPageId: z.string().uuid().nullable(),
})
export type ReorderPageInput = z.infer<typeof reorderPageInput>

export const softDeletePageInput = z.object({
  id: z.string().uuid(),
  workspaceId: z.string().uuid(),
})
export type SoftDeletePageInput = z.infer<typeof softDeletePageInput>

export const restorePageInput = z.object({
  id: z.string().uuid(),
  workspaceId: z.string().uuid(),
})
export type RestorePageInput = z.infer<typeof restorePageInput>

export const hardDeletePageInput = z.object({
  id: z.string().uuid(),
  workspaceId: z.string().uuid(),
})
export type HardDeletePageInput = z.infer<typeof hardDeletePageInput>

export const emptyTrashInput = z.object({
  workspaceId: z.string().uuid(),
})
export type EmptyTrashInput = z.infer<typeof emptyTrashInput>